MRG: write `manysketch` `--singleton` sketches incrementally to reduce memory usage (#375)
I'm running into memory issues with manysketch --singleton, likely brought on by the fact that we're storing all the singleton sigs in a list until the end of the file, rather than streaming + sending them to be written. This PR fixes that by sending singletons for writing right after signature generation (before going on to the next read/contig). We then avoid writing singletons at the end of the loop, where they would just be empty template sigs.